Fixed faulty merge + merged manual.tx and building-manual.ts (#410)
