+++ frontend/components/manual/ManualView.tsx
@@ -1,7 +1,7 @@
-import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
-import { getManual } from "@/lib/manual";
-import { AxiosResponse } from "axios";
+import {useRouter} from "next/router";
+import {useEffect, useState} from "react";
+import {AxiosResponse} from "axios";
+import {getManual} from "@/lib/building-manual";
 
 interface ManualQuery {
     id?: string;
@@ -26,7 +26,9 @@ function ManualView() {
         }
         getManual(Number(query.id))
             .then((manual: AxiosResponse) => {
-                setFile(`${process.env.NEXT_PUBLIC_BASE_API_URL}${manual.data.file.slice(1)}`);
+                const file = `${process.env.NEXT_PUBLIC_BASE_API_URL}${manual.data.file.slice(1)}`;
+                setFile(file);
+                //console.log(file);
             })
             .catch((error) => {
                 //TODO: error component
@@ -48,7 +50,7 @@ function ManualView() {
                         marginTop: ".5%",
                     }}
                 >
-                    <embed src={file} style={{ width: "97%" }} />
+                    <embed src={file} style={{width: "97%"}}/>
                 </div>
             )}
         </>
