+++ frontend/lib/building-manual.ts
@@ -1,6 +1,12 @@
 import api from "@/lib/api/axios";
 import {AxiosResponse} from "axios";
 
+export interface ManualInterface {
+    id: number;
+    building: number;
+    version_number: number;
+    file: string;
+}
 
 export interface BuildingManual {
     id: number,
@@ -14,7 +20,12 @@ export interface BuildingManualPostInterface {
     file: File
 }
 
-export async function getManualsForBuilding(buildingId: number): Promise<AxiosResponse<any>> {
+export const getManual = async (manualId: string | undefined | number): Promise<AxiosResponse<any>> => {
+    const request_url: string = `${process.env.NEXT_PUBLIC_BASE_API_URL}${process.env.NEXT_PUBLIC_API_MANUAL}${manualId}`;
+    return await api.get(request_url);
+}
+
+export async function getManualsForBuilding(buildingId: string | number): Promise<AxiosResponse<any>> {
     const get_url: string = `${process.env.NEXT_PUBLIC_BASE_API_URL}${process.env.NEXT_PUBLIC_API_MANUAL_BUILDING}${buildingId}`;
     return await api.get(get_url, {
         headers: {"Content-Type": "multipart/form-data"},
@@ -23,4 +34,16 @@ export async function getManualsForBuilding(buildingId: number): Promise<AxiosRe
 
 export function getManualPath(relativePath: string) {
     return `${process.env.NEXT_PUBLIC_BASE_API_URL}${relativePath.substring(1)}` // Remove first / of relativePath
+}
+
+export async function postManual(manual: BuildingManualPostInterface): Promise<AxiosResponse<any>> {
+    const formdata = new FormData();
+    formdata.set("building", manual.building.toString());
+    formdata.set("file", manual.file);
+    const request_url: string = `${process.env.NEXT_PUBLIC_BASE_API_URL}${process.env.NEXT_PUBLIC_API_MANUAL}`;
+    return await api.post(request_url, formdata,
+        {
+            headers: {"Content-Type": "multipart/form-data"},
+        }
+    );
 }
